+++ packages/frontend/src/app/(side-nav)/bridges/summary/_components/table/columns.tsx
@@ -1,7 +1,7 @@
 import { createColumnHelper } from '@tanstack/react-table'
 import { TotalCell } from '~/app/(side-nav)/scaling/summary/_components/table/total-cell'
 import { TableValueCell } from '~/components/table/cells/table-value-cell'
-import { TypeCell } from '~/components/table/cells/type-cell'
+import { TypeInfo } from '~/components/table/cells/type-info'
 import {
   adjustTableValue,
   sortTableValues,
@@ -29,7 +29,7 @@ export const bridgesSummaryActiveColumns = [
       tooltip:
         'Token bridges use escrows and mint tokens. Liquidity Networks use pools and swap tokens. Hybrid do both.',
     },
-    cell: (ctx) => <TypeCell>{ctx.getValue()}</TypeCell>,
+    cell: (ctx) => <TypeInfo>{ctx.getValue()}</TypeInfo>,
   }),
   columnHelper.accessor((e) => e.tvs.breakdown?.total, {
     id: 'total',
