+++ packages/frontend/src/app/(side-nav)/scaling/summary/_components/table/columns.tsx
@@ -10,10 +10,11 @@ import {
 import { PizzaRosetteCell } from '~/components/rosette/pizza/pizza-rosette-cell'
 import { StageCell } from '~/components/table/cells/stage/stage-cell'
 import { TableValueCell } from '~/components/table/cells/table-value-cell'
+import { TwoRowCell } from '~/components/table/cells/two-row-cell'
 import {
-  TypeCell,
   TypeExplanationTooltip,
-} from '~/components/table/cells/type-cell'
+  TypeInfo,
+} from '~/components/table/cells/type-info'
 import { ValueWithPercentageChange } from '~/components/table/cells/value-with-percentage-change'
 import { sortStages } from '~/components/table/sorting/sort-stages'
 import { getScalingCommonProjectColumns } from '~/components/table/utils/common-project-columns/scaling-common-project-columns'
@@ -40,13 +41,16 @@ export const scalingSummaryColumns = [
   columnHelper.accessor('category', {
     header: 'Type',
     cell: (ctx) => (
-      <TypeCell
-        stack={ctx.row.original.stack}
-        isAppchain={ctx.row.original.capability === 'appchain'}
-        purposes={ctx.row.original.purposes}
-      >
-        {ctx.getValue()}
-      </TypeCell>
+      <TwoRowCell>
+        <TwoRowCell.First>
+          <TypeInfo stack={ctx.row.original.stack}>{ctx.getValue()}</TypeInfo>
+        </TwoRowCell.First>
+        {ctx.row.original.capability === 'appchain' && (
+          <TwoRowCell.Second>
+            {ctx.row.original.purposes.join(', ')}
+          </TwoRowCell.Second>
+        )}
+      </TwoRowCell>
     ),
     meta: {
       tooltip: <TypeExplanationTooltip />,
